:hammer: small change related to [change sourceId ucTxnId from long to int]

+++ libra-core/src/test/kotlin/xyz/ivyxjc/libra/Commons.kt
@@ -8,7 +8,7 @@ import kotlin.random.Random
 fun buildRawTrans(): RawTransaction {
     val res = RawTransaction()
     res.gcGuid = Random.nextInt().toString()
-    res.sourceId = Random.nextInt().toLong()
+    res.sourceId = Random.nextInt()
     res.rawRecord = RandomStringUtils.random(1000)
     return res
 }
